Revise sjd keyboard after Mac Kildin Sami
Philosophy:
Modified consonant on alt + consonant
Long vowel on alt + vowel
Umlaut vowel next to the long vowel with alt
The two latin letters on alt + ч, й
The obsolete ҋ not under й (since it is not in use), but next to it
The obsolete ʼ under х.
Otherwise, the free positions on the alt and alt + shift level were filled with parentheses, mathematical operators and letters.

The longpress option for mobile phones as expected, but without obsolete letters.

TODO: Look at the cmd and ctrl levels.