All Cyrillik К to the flag
Using Arial Unicode MS in OmniGraffle (or any app able to export PNG with transparent background). Font size set to 1240 pt, then scaled down in the image editor to fit the flag.

The original flag without the letter was taken from the SME keyboard, from the initial commit of the flag.